Added basic locational filtering and heatmap
Added another script that lets you change the filters on the map, and then builds a heatmap of the meteorites in that selected zone.

I don't like the way the heat map looks and I think it isn't displaying correctly? it only shows 0-18 lat and long. When i think it should be -90 to 90 and -180 to 180