Add files via upload
Hi Team, We are Group 2.
For our Final project, the backup dataset can be added as our chapter 2. I explored the dataset and able to identify 11 attributes that could be used to predict the customer satisfaction. 8 of the 11 attributes where categorical so, I encoded them in to numbers. KNN, Decision tree, and SVM did a good job in predicting at 0.99 accuracy. The total rows are 115,000 and I used 80-20 split for training and testing. I have uploaded the jupyter notebook file with code and output. Confusion matrix is also added. Please review.